feat(MonthBar): add month and return bar
